requirements.txt
fastapi
uvicorn
openai
sqlalchemy
psycopg2-binary
faiss-cpu
pandas
requests
python-dotenv
python-docx
pypdf